[src] InputAction -> Action
