+++ Sources/AsyncUI/Action/Action+Void.swift
@@ -1,0 +1,8 @@
+import Foundation
+import Combine
+
+extension Action where Input == Void {    
+    public func callAsFunction() {
+        self.callAsFunction(())
+    }
+}
